changed planning route to work with url search params instead of nested paths to be able to handle more than one class or location of the same type
