+++ src/App.tsx
@@ -128,24 +128,8 @@ const App: FC = () => {
                 <Route path="new" element={<ProjectBasics />} />
                 <Route path="notes" element={<ProjectNotes />} />
               </Route>
-              <Route path="/planning" element={<PlanningView />}>
-                <Route path=":masterClassId" element={<PlanningView />}>
-                  <Route path=":classId" element={<PlanningView />}>
-                    <Route path=":subClassId" element={<PlanningView />}>
-                      <Route path=":districtId" element={<PlanningView />} />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
-              <Route path="/coordination" element={<PlanningView />}>
-                <Route path=":masterClassId" element={<PlanningView />}>
-                  <Route path=":classId" element={<PlanningView />}>
-                    <Route path=":subClassId" element={<PlanningView />}>
-                      <Route path=":districtId" element={<PlanningView />} />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
+              <Route path="/planning" element={<PlanningView />} />
+              <Route path="/coordination" element={<PlanningView />} />
               <Route path="/search-results" element={<SearchResultsView />} />
               <Route path="*" element={<ErrorView />} />
             </Routes>
